requirements.txt
streamlit>=1.38.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-chroma>=0.2.0
chromadb>=0.5.0
PyMuPDF>=1.24.0
python-docx>=1.1.0
Pillow>=10.0.0
